arch/arm64: the arm64 perf interface supports pmu
Summary:
- Support arm64 pmu api, Currently only the cycle counter function is supported.
- Using ARM64 PMU hardware capability to implement perf interface, modify all
  perf interface related code.
- Support for pmu init under smp.

+++ arch/arm/src/armv7-m/arm_perf.c
@@ -34,7 +34,7 @@
  * Private Data
  ****************************************************************************/
 
-static uint32_t g_cpu_freq;
+static unsigned long g_cpu_freq;
 
 /****************************************************************************
  * Public Functions
@@ -42,7 +42,7 @@ static uint32_t g_cpu_freq;
 
 void up_perf_init(void *arg)
 {
-  g_cpu_freq = (uint32_t)(uintptr_t)arg;
+  g_cpu_freq = (unsigned long)(uintptr_t)arg;
 
   /* Enable ITM and DWT resources, if not left enabled by debugger. */
 
@@ -56,19 +56,19 @@ void up_perf_init(void *arg)
   modifyreg32(DWT_CTRL, 0, DWT_CTRL_CYCCNTENA_MASK);
 }
 
-uint32_t up_perf_getfreq(void)
+unsigned long up_perf_getfreq(void)
 {
   return g_cpu_freq;
 }
 
-uint32_t up_perf_gettime(void)
+unsigned long up_perf_gettime(void)
 {
   return getreg32(DWT_CYCCNT);
 }
 
-void up_perf_convert(uint32_t elapsed, struct timespec *ts)
+void up_perf_convert(unsigned long elapsed, struct timespec *ts)
 {
-  uint32_t left;
+  unsigned long left;
 
   ts->tv_sec  = elapsed / g_cpu_freq;
   left        = elapsed - ts->tv_sec * g_cpu_freq;
