+++ sched/sched/sched_critmonitor.c
@@ -103,8 +103,8 @@
 
 /* Start time when pre-emption disabled or critical section entered. */
 
-static uint32_t g_premp_start[CONFIG_SMP_NCPUS];
-static uint32_t g_crit_start[CONFIG_SMP_NCPUS];
+static unsigned long g_premp_start[CONFIG_SMP_NCPUS];
+static unsigned long g_crit_start[CONFIG_SMP_NCPUS];
 
 /****************************************************************************
  * Public Data
@@ -112,8 +112,8 @@ static uint32_t g_crit_start[CONFIG_SMP_NCPUS];
 
 /* Maximum time with pre-emption disabled or within critical section. */
 
-uint32_t g_premp_max[CONFIG_SMP_NCPUS];
-uint32_t g_crit_max[CONFIG_SMP_NCPUS];
+unsigned long g_premp_max[CONFIG_SMP_NCPUS];
+unsigned long g_crit_max[CONFIG_SMP_NCPUS];
 
 /****************************************************************************
  * Public Functions
@@ -148,8 +148,8 @@ void nxsched_critmon_preemption(FAR struct tcb_s *tcb, bool state)
     {
       /* Re-enabling.. Check for the max elapsed time */
 
-      uint32_t now     = up_perf_gettime();
-      uint32_t elapsed = now - tcb->premp_start;
+      unsigned long now     = up_perf_gettime();
+      unsigned long elapsed = now - tcb->premp_start;
 
       if (elapsed > tcb->premp_max)
         {
@@ -196,8 +196,8 @@ void nxsched_critmon_csection(FAR struct tcb_s *tcb, bool state)
     {
       /* Leaving .. Check for the max elapsed time */
 
-      uint32_t now     = up_perf_gettime();
-      uint32_t elapsed = now - tcb->crit_start;
+      unsigned long now     = up_perf_gettime();
+      unsigned long elapsed = now - tcb->crit_start;
 
       if (elapsed > tcb->crit_max)
         {
@@ -230,9 +230,9 @@ void nxsched_critmon_csection(FAR struct tcb_s *tcb, bool state)
 
 void nxsched_resume_critmon(FAR struct tcb_s *tcb)
 {
-  uint32_t current = up_perf_gettime();
+  unsigned long current = up_perf_gettime();
   int cpu = this_cpu();
-  uint32_t elapsed;
+  unsigned long elapsed;
 
   tcb->run_start = current;
 
@@ -294,8 +294,8 @@ void nxsched_resume_critmon(FAR struct tcb_s *tcb)
 
 void nxsched_suspend_critmon(FAR struct tcb_s *tcb)
 {
-  uint32_t current = up_perf_gettime();
-  uint32_t elapsed = current - tcb->run_start;
+  unsigned long current = up_perf_gettime();
+  unsigned long elapsed = current - tcb->run_start;
 
   if (elapsed > tcb->run_max)
     {
